requirements/base.txt
# Core Django
Django==5.0.7
djangorestframework==3.15.2
django-environ==0.11.2

# Database
psycopg[binary]==3.2.1
dj-database-url==2.2.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
dj-rest-auth==6.0.0

# CORS
django-cors-headers==4.4.0

# Storage (AWS S3)
django-storages==1.14.4
boto3==1.34.144

# Image Processing
Pillow==10.4.0

# Background Tasks
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.27.2

# Utilities
python-decouple==3.8
pytz==2024.1

